+++ src/main/kotlin/de/mariushoefler/flutterenhancementsuite/utils/FlutterProjectUtils.kt
@@ -2,65 +2,33 @@ package de.mariushoefler.flutterenhancementsuite.utils
 
 import com.intellij.openapi.fileEditor.FileDocumentManager
 import com.intellij.openapi.project.Project
-import com.intellij.openapi.vfs.VfsUtil
-import com.intellij.openapi.vfs.VirtualFile
+import com.intellij.psi.PsiDocumentManager
+import com.intellij.psi.PsiFile
 import io.flutter.pub.PubRoot
 import io.flutter.sdk.FlutterSdk
-import org.yaml.snakeyaml.DumperOptions
-import org.yaml.snakeyaml.Yaml
-import org.yaml.snakeyaml.constructor.SafeConstructor
-import org.yaml.snakeyaml.nodes.Tag
-import org.yaml.snakeyaml.representer.Representer
-import org.yaml.snakeyaml.resolver.Resolver
-import java.io.File
-import java.io.IOException
 
 object FlutterProjectUtils {
-
-    fun readProjectName(project: Project): String? {
-        val pubspec =
-            VfsUtil.findFileByIoFile(File("${project.basePath}${File.separator}pubspec.yaml"), true) ?: return null
-        val properties = readPubspecFileToMap(pubspec)
-        return properties?.get("name") as String
-    }
-
-    @Throws(IOException::class)
-    fun readPubspecFileToMap(pubspec: VirtualFile): Map<String, Any>? {
-        val contents = String(pubspec.contentsToByteArray(true))
-        return loadPubspecInfo(contents)
-    }
-
-    private fun loadPubspecInfo(yamlContents: String): Map<String, Any>? {
-        val yaml = Yaml(
-            SafeConstructor(),
-            Representer(),
-            DumperOptions(),
-            object : Resolver() {
-                override fun addImplicitResolvers() {
-                    this.addImplicitResolver(Tag.BOOL, BOOL, "yYnNtTfFoO")
-                    this.addImplicitResolver(Tag.NULL, NULL, "~nN\u0000")
-                    this.addImplicitResolver(Tag.NULL, EMPTY, null)
-                    this.addImplicitResolver(Tag("tag:yaml.org,2002:value"), VALUE, "=")
-                    this.addImplicitResolver(Tag.MERGE, MERGE, "<")
-                }
-            }
-        )
-
-        return yaml.load<Map<String, Any>>(yamlContents)
-    }
-
     /**
      * Runs `flutter pub get` in project
      *
      * @since v1.2
      */
-    fun runPackagesGet(file: VirtualFile, project: Project) {
-        PubRoot.forDescendant(file, project)?.let { pubRoot ->
-            FileDocumentManager.getInstance().saveAllDocuments()
-            pubRoot.getModule(project)?.let { module ->
-                FlutterSdk.getFlutterSdk(project)?.flutterPackagesGet(pubRoot)
-                    ?.startInModuleConsole(module, { pubRoot.refresh() }, null)
+    fun runPackagesGet(file: PsiFile, project: Project) {
+        PubRoot.forDescendant(file.virtualFile, project)?.let { pubRoot ->
+            PsiDocumentManager.getInstance(project)?.let {
+                it.getDocument(file)?.let { doc ->
+                    it.doPostponedOperationsAndUnblockDocument(doc)
+                    FileDocumentManager.getInstance().saveAllDocuments()
+                    executePubGet(pubRoot, project)
+                }
             }
+        }
+    }
+
+    private fun executePubGet(pubRoot: PubRoot, project: Project) {
+        pubRoot.getModule(project)?.let { module ->
+            FlutterSdk.getFlutterSdk(project)?.flutterPackagesGet(pubRoot)
+                ?.startInModuleConsole(module, { pubRoot.refresh() }, null)
         }
     }
 }