+++ CHANGELOG.md
@@ -2,9 +2,13 @@
 
 ## [Unreleased]
 ### Added
+- Open a package's pub.dev page directly from your pubspec.yaml (#121)
 - Custom file templates for "New Flutter widget" menu
+- Run `flutter pub get` from menu by right-clicking a file or a directory (#52)
 ### Changed
 - Const constructor added for StatelessWidget and StatefulWidget templates (#141)
+### Fixed
+- Fix "Document is locked by write PSI operations" (#124)
 ### Removed
 - Remove coverage feature
 - Remove bloc feature